Started testing SP -> TCv4 excel conversion. Fixed Skills TCv4 mysh parsing. Fixed crash bug if TCv4 excel files not found. Fixed Options display for TCv4 data files not showing correct state. Fixed .wav and .wm multimedia types being patched out. Fixed .dds lightingmap being patched out - appears that texture .dds files don't matter. (make this more specific?) Fixed Excel BackGroundSounds2D class types.

+++ hellgate/Excel/BackGroundSounds2D.cs
@@ -12,8 +12,8 @@ namespace Hellgate.Excel
         [ExcelOutput(SortColumnOrder = 1)]
         [MarshalAs(UnmanagedType.ByValTStr, SizeConst = 64)]
         public string name;
-        public float undefined1;
-        public float undefined2;
+        public Int32 undefined1;
+        public Int32 undefined2;
         public float minVolume;
         public float maxVolume;
         public float silentChance;
